🚚 Rename `src/layout/` folder to use plural form
Every other folder of this kind in the app uses the plural form, so this one should too, for consistency.
