+++ src/layouts/AuthenticatedLayout/style.scss
