+++ src/layouts/SimplePageLayout/style.scss
