+++ src/views/ActivityFeed/index.jsx
@@ -1,6 +1,6 @@
 import { ScrollTopButton } from '../../components/Buttons/ScrollTopButton'
 import { Feed } from '../../components/Feed'
-import { AuthenticatedLayout } from '../../layout/AuthenticatedLayout'
+import { AuthenticatedLayout } from '../../layouts/AuthenticatedLayout'
 
 import './style.scss'
