+++ src/views/Administration/index.jsx
@@ -1,4 +1,4 @@
-import { AuthenticatedLayout } from '../../layout/AuthenticatedLayout'
+import { AuthenticatedLayout } from '../../layouts/AuthenticatedLayout'
 import { AdminMembers } from '../../components/Admin/AdminMembers'
 import { ScrollTopButton } from '../../components/Buttons/ScrollTopButton'
 import './style.scss'
