+++ src/views/OrganizationCreation/index.jsx
@@ -1,5 +1,5 @@
 import { OrganizationForm } from '../../components/Forms/OrganizationForm'
-import { DoublePageLayout } from '../../layout/DoublePageLayout'
+import { DoublePageLayout } from '../../layouts/DoublePageLayout'
 
 import './style.scss'
