+++ src/views/SignUp/index.jsx
@@ -1,5 +1,5 @@
 import { ProfileForm } from '../../components/Forms/ProfileForm'
-import { DoublePageLayout } from '../../layout/DoublePageLayout'
+import { DoublePageLayout } from '../../layouts/DoublePageLayout'
 
 import './style.scss'
